CA| Adding bednet client fields
I focused on bednet form for this commit. I realized there were a few fields that needed to be updated. I also added the ec_client fields and classification as well as the bednet to the ListTaskInteractor. I added constants for all three forms as well so they were easier to reference.